Month 05

+++ database/migrations/2021_09_05_165054_create_months_table.php
@@ -18,7 +18,7 @@ class CreateMonthsTable extends Migration
       $table->smallInteger('month');
       $table->unsignedBigInteger('user_id');
       $table->mediumInteger('bruto')->nullable();
-      $table->smallInteger('prijevoz')->nullable();
+      $table->mediumInteger('prijevoz')->nullable();
       $table->mediumInteger('odbitak')->nullable();
       $table->smallInteger('prirez')->nullable();
       $table->tinyInteger('prekovremeni')->nullable();
